+++ gtfs_parser/constants.py
